+++ CHANGES.rst
@@ -4,13 +4,13 @@ Changelog
 1.2 (unreleased)
 ----------------
 
-- Add support for Python 3.8.
-
-- Drop support for Python 3.4.
+- Add support for Python 3.8, drop support for Python 3.4.
 
 - Improve error message in case zc.buildout is not installed.
 
 - Improve installation instruction.
+
+- Pass the index url to both the 'installed' and 'buildout' checkers.
 
 
 1.1 (2018-11-03)